Use floats for animation scaling.
Vision boundaries look reasonable, and the sprites themselves are still slightly
bigger than the player.

+++ include/Actor.h
@@ -112,7 +112,7 @@ class Actor {
 		sf::Clock sprite_clock;
 		unsigned frame_duration;
 		unsigned frame_no;
-		unsigned animation_scaling;
+		float animation_scaling;
 
 		void animate();
 };
